backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
social-auth-app-django
django-otp
django-two-factor-auth
phonenumbers
pyotp
django_extensions
Werkzeug
pyOpenSSl
Pillow
qrcode
requests

